src/v3n0m.py: - Document recovery menu + declare its hacks
Forced upgrade is against my moral ode but sometimes it's more consistent.
It could be worse. But still, don't run this as root, please.

You can Ctrl^C to escape the shell but it will hang and possibly segfault.
I have tried to optimize it but nothing I do has fixed it.
System upgrades all dependencies each run and it's sort of time-consuming.
Unstable, as is.

cachestatus():
- This is what you want to edit to show more/less of the data collected.
By default it only shows you the counters, but you can display what sites
specifically if desired, just print the value to stdout.