Slim down Dockerfile and apply deps upgrades
This is a combo patch for both the Dockerfile and the Poetry lockfile.
Simplify Poetry installation in Docker image.

* docker/Dockerfile:

 set `ARG PYTHON_VERSION` to 3.10
 use `${PYTHON_VERSION}-slim-bullseye` as builder from now on
 cleanup `RUN` where we install dependencies for the image
 cleanup `RUN` where we install Poetry:
 update to Poetry 1.2.1 from 1.1.14
 keep magic - we like magic - but simplify the installation

This should make installation easier on everyone by explicitly stating
cache location for pip and installing as, yes, root. Installing as user
`venom` is too complicated and this Dockerfile *should* be safe as-is
even running as root. There are still ways to harden to image that don't
require huge lines to add a user environment and it's less hacky this way.

* docker/README:

 specify `Yggdrasil` as tag
 docker trust no longer functions as expected; remove instructions
 building as newly defined tag still works, pull from the hub or build local

* poetry.lock:

  Package operations: 0 installs, 8 updates, 0 removals

  • Updating idna (3.3 -> 3.4)
  • Updating attrs (21.4.0 -> 22.1.0)
  • Updating certifi (2022.6.15 -> 2022.9.24)
  • Updating pytz (2022.1 -> 2022.4)
  • Updating urllib3 (1.26.11 -> 1.26.12)
  • Updating aiohttp (3.8.1 -> 3.8.3)
  • Updating datetime (4.5 -> 4.7)
  • Updating tqdm (4.64.0 -> 4.64.1)

Signed-off-by: vittring <scissortail@riseup.net>